refacted code
Condensed code into a more pure form.

Nested global varibles in a IIFE.

Created an add and getAll function within the IIFE to add and retrive pokemons.

The IIFE's return statement returns an object with functions that are avaible inside IIFE. (Returning object is store in pokemonRepository)

Used to forEach function of the pokemonRepository object to code functions within IIFE.